type checking implemented

+++ src/main/java/ca/utoronto/utm/mcs/Post.java
@@ -13,15 +13,11 @@ import org.json.*;
 
 import com.mongodb.client.MongoClient;
 import com.mongodb.BasicDBObject;
-import com.mongodb.DBObject;
-import com.mongodb.client.ClientSession;
 import com.mongodb.client.FindIterable;
-import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoCursor;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.result.DeleteResult;
-import com.mongodb.session.ServerSession;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 
@@ -79,16 +75,20 @@ public class Post implements HttpHandler, AutoCloseable
 	 		return;
 	 	}
 	 	
-	String title;
+	 	String title;
         String author;
         String content;
         List<String> tags = new ArrayList<String>();
         
-        if (deserialized.has("title") && deserialized.has("author") 
-        		&& deserialized.has("content") && deserialized.has("tags")) {
-            title = deserialized.getString("title");
-            author = deserialized.getString("author");
-            content = deserialized.getString("content");
+        if (deserialized.has("title") && deserialized.get("title") instanceof String 
+        		&& deserialized.has("author") && deserialized.get("author") instanceof String 
+        		&& deserialized.has("content") && deserialized.get("content") instanceof String 
+        		&& deserialized.has("tags")) {
+
+        	title = deserialized.getString("title");
+        	author = deserialized.getString("author");
+        	content = deserialized.getString("content");
+        	
             try {
             	JSONArray arr = deserialized.getJSONArray("tags");
             	for(int i = 0; i < arr.length(); i++) {
@@ -103,13 +103,12 @@ public class Post implements HttpHandler, AutoCloseable
         	r.sendResponseHeaders(400, -1);
         	return;
         }
-	// Good so connect to mongo and post
+        // Good so connect to mongo and post
         JSONObject post = new JSONObject();
         post.put("title", title);
         post.put("author", author);
         post.put("content", content);
         post.put("tags", tags);
-        	
         Document doc = Document.parse(post.toString());
         posts.insertOne(doc);
         ObjectId id = doc.getObjectId("_id");
@@ -123,8 +122,8 @@ public class Post implements HttpHandler, AutoCloseable
 	
 	} catch (Exception e) {
 		 //IF IT ever errors out i guess 500 takes priority
-		 r.sendResponseHeaders(500, -1);
-		 return;
+		r.sendResponseHeaders(500, -1);
+		return;
 	}
 
 }
@@ -143,47 +142,46 @@ public void handleGet(HttpExchange r) throws IOException, JSONException {
         String id;
         String title;
         String contents = "["; //start of response
-        
    
         //id only, or id takes priority --------------------------------------------------------------------------------------------
         if (deserialized.has("_id")) {
             id = deserialized.getString("_id");
-            
+            if (!ObjectId.isValid(id)) {//has both but id is incorrect RETURN 400 according to ilir )
+            	r.sendResponseHeaders(400, -1);
+            	return;
+            }
             //query for _id = id
             BasicDBObject whereQuery = new BasicDBObject();
             whereQuery.put("_id", new ObjectId(id));
             FindIterable<Document> cursor = posts.find(whereQuery);
           
-            if (cursor.first() == null && deserialized.has("title")) //has both but id is incorrect RETURN 400 according to ilir 
-            	r.sendResponseHeaders(400,-1);
-            else if (cursor.first() == null) //otherwise
+            if (cursor.first() == null) { 
             	r.sendResponseHeaders(404,-1);
+            	return;
+            }
             Document d = cursor.first();
-            
-            System.out.println(d.get("_id"));
-            if (!ObjectId.isValid(id))
-            	r.sendResponseHeaders(400, -1);
             contents += this.generate_response(d);
         }
         //title only ----------------------------------------------------------------------------------------------------------------
-        else if (deserialized.has("title")) {
+        else if (deserialized.has("title") && deserialized.get("title") instanceof String) {
         	title = deserialized.getString("title");
-        	
         	//query for all titles that contain string "title"
             BasicDBObject regexQuery = new BasicDBObject();
             regexQuery.put("title", 
                 new BasicDBObject("$regex", title));
             FindIterable<Document> cursor = posts.find(regexQuery);
-            if (cursor.first() == null) //no post(s) found
+            if (cursor.first() == null) {//no post(s) found
             	r.sendResponseHeaders(404,-1);
-            
+            	return;
+            }
             Iterator<Document> T = cursor.iterator();
             while (T.hasNext()) {
             	Document D = T.next();
             	String temp_Id = D.getObjectId("_id").toString();
-            	if (!ObjectId.isValid(temp_Id))
+            	if (!ObjectId.isValid(temp_Id)) {
                 	r.sendResponseHeaders(400, -1);
-            	
+                	return;
+            	}
             	contents += this.generate_response(D);
             	if (T.hasNext()) //if NOT last doc then add comma
             		contents += ",";
@@ -196,15 +194,7 @@ public void handleGet(HttpExchange r) throws IOException, JSONException {
         }     
         
         contents += "\n]"; //end of response
-
-        System.out.println(contents);
         
-//        JSONObject test = new JSONObject();
-//        test.put("title", "abc");
-//        test.put("author", "123");
-//        test.put("content", "321");
-//        test.put("tags", "lul");
-//        
         r.sendResponseHeaders(200, contents.length()); //response.length());
         OutputStream os = r.getResponseBody();
         os.write(contents.getBytes());		//response.getBytes());
@@ -273,13 +263,17 @@ public void handleDelete(HttpExchange r) throws IOException, JSONException {
 		}
 
 		String Id;
-
+		
 		if (deserialized.has("_id"))
 			Id = deserialized.getString("_id");
 		else {
 			r.sendResponseHeaders(400, -1);
 			return;
 		}
+		if (!ObjectId.isValid(Id)) {//has both but id is incorrect RETURN 400 according to ilir )
+        	r.sendResponseHeaders(400, -1);
+        	return;
+        }
 		Document doc = Document.parse("{\"_id\": ObjectId(\"" + Id + "\")}");
 		DeleteResult res = posts.deleteOne(doc);
 		if(res.getDeletedCount() == 0) {
